[~] geometry: Make index buffer use unsigned short

+++ src/core/include/ttb/core/geometry/IndexBuffer.hpp
@@ -14,7 +14,7 @@ namespace ttb
     class IndexBuffer
     {
     public:
-        using Index = GLuint;
+        using Index = GLushort;
 
         static std::shared_ptr< IndexBuffer > create();
